url-shortener-fastapi/backend/pyproject.toml
[project]
name = "url-shortener-fastapi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.16.2",
    "dotenv>=0.9.9",
    "fastapi[standard]>=0.115.14",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2-binary>=2.9.10",
    "python-jose[cryptography]>=3.5.0",
    "sqlalchemy>=2.0.41",
]
